+++ tests/test_integration_unix.py
@@ -5,9 +5,9 @@ import pytest
 import aiospamc
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
+pytestmark = pytest.mark.skipif(sys.platform == "win32", reason="Unix sockets not supported on Windows")
+
+
 @pytest.mark.integration
 async def test_check(spamd_unix, spam):
     result = await aiospamc.check(spam, socket_path=spamd_unix)
@@ -15,9 +15,6 @@ async def test_check(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_headers(spamd_unix, spam):
     result = await aiospamc.headers(spam, socket_path=spamd_unix)
@@ -25,9 +22,6 @@ async def test_headers(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_ping(spamd_unix):
     result = await aiospamc.ping(socket_path=spamd_unix)
@@ -35,9 +29,6 @@ async def test_ping(spamd_unix):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_process(spamd_unix, spam):
     result = await aiospamc.process(spam, socket_path=spamd_unix)
@@ -45,9 +36,6 @@ async def test_process(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_report(spamd_unix, spam):
     result = await aiospamc.report(spam, socket_path=spamd_unix)
@@ -55,9 +43,6 @@ async def test_report(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_report_if_spam(spamd_unix, spam):
     result = await aiospamc.report_if_spam(spam, socket_path=spamd_unix)
@@ -65,9 +50,6 @@ async def test_report_if_spam(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_symbols(spamd_unix, spam):
     result = await aiospamc.symbols(spam, socket_path=spamd_unix)
@@ -75,9 +57,6 @@ async def test_symbols(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_tell(spamd_unix, spam):
     result = await aiospamc.tell(
@@ -87,9 +66,6 @@ async def test_tell(spamd_unix, spam):
     assert 0 == result.status_code
 
 
-@pytest.mark.skipif(
-    sys.platform == "win32", reason="Unix sockets not supported on Windows"
-)
 @pytest.mark.integration
 async def test_message_without_newline(spamd_unix):
     result = await aiospamc.check(message=b"acb", socket_path=spamd_unix)
